knowledge-base/requirements.txt
# RAG knowledge base for NANDA Work
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
